requirements (2).txt
pyrogram
tgcrypto
python-dotenv
flask
uvloop
